working on linux

+++ test/common/execution_thread_pool_test.cpp
@@ -232,13 +232,13 @@ TEST(ExecutionThreadPoolTests, TaskStealingCorrectnessTest) {
 
     std::atomic<int16_t> order_count = 0;
     common::SpinLatch order_latch_;
-    std::map<uint32_t, storage::numa_region_t> numa_order;
+    std::map<int16_t, storage::numa_region_t> numa_order;
     std::promise<void> check_promises[num_numa_regions];
     for (int16_t i = 0; i < num_numa_regions; i++) {
       storage::numa_region_t numa_hint = static_cast<storage::numa_region_t>(i);
       auto workload = [&, numa_hint] {
         common::SpinLatch::ScopedSpinLatch l(&order_latch_);
-        uint16_t pos = order_count++;
+        int16_t pos = order_count++;
         TERRIER_ASSERT(numa_order.find(pos) == numa_order.end(), "there should be no other node at this position");
         numa_order[pos] = numa_hint;
       };
